Fix concourse terraform state version pinning.
It's important that every terraform run uses the latest state file so
that it has an accurate reflection of what's been created etc.

Previously, the cf.tfstate used in the deploy-cloudfoundry pipeline was
pinned to the version that has passed the init step. This meant that
if a terraform run failed for any reason (eg AWS rate limit), and the
job was restarted, it would not use the updated statefile that included
the changes made in the previous run.

This changes the pipeline to use a non-pinned version of the statefile
to ensure it always uses the latest.

It also stops the init job from `put`ing the resource (leaving
s3init.sh to do the writing). This makes this pipeline consistent with
the others.