Refactor CF cert generation to use single tarball
Instead of having a concourse resource for each certificate, this uses a
single tarball containing all the certificates used in the cloudfoundry
manifest. This makes the pipeline simpler and clearer. This also removes
some of the more complex usage of the s3init script.

This also includes improvements to the log output from these tasks.