Use correct domains in router SSL cert.
This was previously using a hardcoded domain for the router SSL cert.
This updates this to get the domains to use from the terraform outputs.
This ensures that it will match what cloudfoundry is configured with.

This also updates the cert to have entries for both the system domain
and the apps domain. This is necessary while there are system components
that are served by the router.