Group network_acls by subnet and direction
The rule ID, subnet, and direction of these network ACLs are all
significant. Having them in separate files made it quite hard to comprehend
the effect on a single subnet. I'm hoping that by grouping them together
like this it is easier to read.

I've made no changes to the actual resources. Although that's quite hard to
verify without parsing and normalising their order.