Extract CF certificate generation into script
This means it can be run through shellcheck. It also removes the need
for the interpolation workarounds.

This includes some refactoring of the script to make it simpler. I think
the orginal version was written to allow for multiple CAs, which added
to the complexity. As we're using a single CA, this can be simplified.