Enable ssl validation for smoke and acceptance tests.
Now that the router certificates have the correct hostnames in them, we
can enable SSL validation in the tests. We just need the bosh-CA cert to
be trusted by the machine/container running the tests.

The smoke tests, and upstream acceptance tests run as errands, therefore
bosh is already adding the CA cert.

The custom acceprance tests run in a container in concourse, so we need
to add the bosh-CA to its root certificates.