Move generate-*-{secrets,certs} scripts
To new `scripts/` directories under their respective `manifests/`
directories. I want to write spec tests for two of these scripts, which is
easier if they are located near the same directory as the tests. It also
feels more logical - since the contents is more closely related with the
BOSH or CF manifest rather than the Concourse pipeline.