README: Add section about test dependencies
In order to run the tests we need quite a few dependencies installed. These
are the sum of several Docker images that we use in our Concourse pipelines.

We should make it possible for people to run the tests locally on their own
machines, but I don't want to duplicate all of the information about what
they are, what versions, and how to install - because it will quickly become
out-of-date.

As a compromise I'm adding a section to the README to point them at the
Travis configuration which is our most canonical reference.