Lint all static YAML files for duplicate keys
The YAML spec is a little bit vague about the exact behaviour a parser
should implement when it encounters duplicate keys in a document. For
example, the go-yaml library that spruce uses takes the last value and
doesn't raise any warnings or errors. This can cause expected results if we
make a copy/paste mistake in a document

Since we have some very large BOSH and Concourse configurations, it seems
quite likely that this could happen again. Prevent it, by running `yamllint`
over all static YAML documents with a configuration file that will only
error on duplicate keys.

We have to install `yamllint` using `pip` to make it available to Travis.
The `--user` argument is necessary because we're running on Travis's
container-based infrastructure which doesn't allow root access:

```
    error: could not create '/usr/local/lib/python2.7/dist-packages/yaml': Permission denied

    ----------------------------------------
    Command "/usr/bin/python -c "import setuptools, tokenize;__file__='/tmp/pip-build-heykkt/pyyaml/setup.py';exec(compile(getattr(tokenize, 'open', open)(__file__).read().replace('\r\n', '\n'), __file__, 'exec'))" install --record /tmp/pip-qEgumI-record/install-record.txt --single-version-externally-managed --compile" failed with error code 1 in /tmp/pip-build-heykkt/pyyaml

The command "pip install yamllint" failed and exited with 1 during .

Your build has been stopped.
```

The preceding commit that updates the README should help people understand
what they need to install locally. I've parameterised it in the same way as
`shellcheck` so you could run it under Docker if you wish.

We can configure other yamllint options in the future, if we wish.