Disable generate-cf-secrets test on Mac
This test fails on Mac because there isn't an `mkpasswd` command available.
Short of writing our own hash+salt function, it seems there is no portable
alternative. So we have to mark this test as "pending" if being run on Mac.

I think that because the tests are run using Bundler that the `rubygems`
package, which provides us with this convenient OS string, will always be
present.