Test and lint secret generation scripts
The YAML spec is a little bit vague about the exact behaviour a parser
should implement when it encounters duplicate keys in a document. For
example, the go-yaml library that spruce uses takes the last value and
doesn't raise any warnings or errors. This can cause expected results if we
make a copy/paste mistake in a document

We are already linting static YAML documents in a preceding commit, but that
doesn't catch the documents outputted by these scripts which generate
passwords on-the-fly. In order to test these, we call them with rspec and
pass the output on to `yamllint`, then check the exit code.

In order to run these scripts on Travis we need to install the packages:

- `pwgen`: which provides the `pwgen` command used to generate passwords
- `whois`: which provides the `mkpasswd` command (go figure - added comment
  to clarify) to hash passwords

The `mkpasswd` is subtly different from the version provided by busybox
which was use in Alpine Linux. Unfortunately the Ubuntu build of busybox is
too old to include the `mkpasswd` binary and `cryptpw`, which is
functionality equivalent, is disabled:

- http://bazaar.launchpad.net/~ubuntu-branches/ubuntu/precise/busybox/precise/view/head:/debian/config/pkg/static#L462

It's not possible to run the `generate-cf-secrets.sh` test on Mac because
there is no `mkpasswd` command available. I'll try to address this in a
following commit.

There is notable duplication between the two tests here. I'm going to
suggest that we raise a separate story to make a common set of rspec
helpers, that are shared between the BOSH and CF tests.