Increase concourse ELB idle timeout
The default is a timeout of 60 seconds. This is problematic because it
causes `fly intercept` sessions to be aborted after 60 seconds of
inactivity.

Increasing this to 600 seconds (10 mins) is much more reasonable for
this use case. In future we plan to remove the concourse ELB entirely,
and access directly using the EIP.