Use new bosh-cli container everywhere.
Now that we have this container available, it makes sense to use it
everywhere we need the bosh CLI. This avoids using the bosh resouece
container, which isn't intended to be used in this way.

This required a fix to the bosh_ensure_uploaded.sh script as it was
using a bashism, but had /bin/sh in the #! line. In the new container,
/bin/sh is no longer bash, so this was erroring.