Extract bosh_login script.
We are awking the bosh password from the yml and using it to login in
several places in the pipelines. It therefore makes sense to extract
this into a script to reduce the duplication.

As part of this the login steps have been removed from
bosh_ensure_uploaded.sh. It now relies on bosh_login.sh having been
called before it's called.