Optionally trigger builds from paas-cf master
This will cause builds to automatically trigger from paas-cf if the
environment variable ENABLE_AUTO_DEPLOY=1 is set. We can be sure that
the tests have already passed as we have master protection turned on, so
the Travis build must have already passed.

Sadly we have to do this in a slighty hacky manner as concourse
currently forcibly quotes all interpolated variables, making them
strings in the eyes of yaml. There's an open issue for this:
https://github.com/concourse/fly/issues/53

Ideally concourse would fix this, however we could also work around this
by using spruce to merge the variables into the pipeline before passing
the newly merged output to `fly` rather than using a separate variables
file.

If we find ourselves having to use this hacky fix more often we should
revisit the spruce alternative solution.

P.S. Strings in the Eyes of Yaml is my new prog-rock band name.