Disable "Adding a wildcard route to a domain" test
After we started to generate SSL certs for our CF deployment and
switched on SSL verification for acceptance tests, "Adding a wildcard
route to a domain" test fails due to a improper curl call that does not
skip SSL verification for multilevel subdomains.

This is fixed in CF release 230. Details can be found here:
https://www.pivotaltracker.com/n/projects/1358110/stories/105340048