Specify release versions etc in the manifest.
Newer versions of the bosh CLI automatically check and upload releases
from the manifest when deploying[1]. This means that it's no longer
necessary for us to do this manually.

This has the added advantage of specifying all the details about the
releases we're using in one place.

The downside of this change is that it no longer parallelises the
downloads for a new deployment, but I consider that cost worth it in
exchange for the simplification here.

[1] https://www.pivotaltracker.com/story/show/102538456