Add bosh-template-renderer code by Simon Johansson
Include the code of https://github.com/simonjohansson/bosh-template-renderer,
a small ruby script which allows render a job template with the spec and
manifest properties.

This will be used to generate the templates provided in releases without
deploying with bosh. For instance, to run the smoketests without using
an errand.