Run smoketests in a concourse instead of errand
Running the tests in errands is quite slow, which can be a problem
when testing the time to recover.

Moving these tests to run inside of concourse will have several
benefits: the execution will be quicker and reduces the feedback loop,
it is easier to troubleshoot, etc.

To run the smoke tests errand we render the original templates from the
cf-release repository for the smoke-tests jobs. This way we can track
and include any changes in the configuration in future releases.

As we use the same script to run the smoke tests than the one used in
the errand, we must create some links to reproduce the same directory
structure as expected in the errand VM.

Also, as we run the job from concourse, we need to install and trust
the bosh CA certificate, which is trusted by default in the errand.