Add logsearch components
We put all logsearch components into a separate file. We have decided
to merge this into main CF deployment for the sake of simplicity. Else
we would have to adjust manifest generation scripts and concourse
deployment pipelines to support multiple deployments, which was out
of scope of this story.

The source of this manifest is logsearch bosh release:

https://github.com/logsearch/logsearch-boshrelease/tree/e58ebfbafd8e0de075a26f78b0b26d3267f083cc/templates

We have combined multiple files they use into a single manifest that
integrates well with our deployment.