Disable require_https for uaa.
The default for this value changed from false to true in newer releases
- see https://docs.google.com/a/pivotal.io/document/d/1GeEvfACj33LS1Ty-yeHx4DPbxvobCjLUnnoexh-wYP8/edit?usp=sharing
(linked from https://github.com/cloudfoundry/cf-release/releases/tag/v227)

When this is set to true, uaa requires all inbound requests to be over
https, and will redirect non-https requests. We are terminating SSL for
these requests in the gorouter, meaning that the requests arrive at UAA
over http. UAA seems to ignore the X-Forwardrd-Proto header, and
therefore this results in a redirect loop.

At some point we'll want to enable SSL for the connection between
gorouter and UAA, at which point this can be set to true.