Fix triggers in deploy-cloudfoundry pipeline
We have noticed that pipeline triggering mechanism is not consistent
across deploy pipeline and that terraform job is not triggering
generate-cf-certs job. In this commit we add pipeline-trigger semver
resource across entire pipeline to fix this problem.

Please note that this update has a side effect of triggering every job
that received pipeline-trigger during pipeline upload. It will happen
only for the first time. New trigger
resource is checking for new version and finds it in 'init' job so it
starts.