Split CF networks to be able to add a SG for RDS
The BOSH CPI implementation only allows define security in the network
definition, although the AWS api does not have that restriction.

Ideally, the BOSH AWS CPI should allow to define the security group per
resource pool, instead of per network.

As we want to add an additional security group to the API and UAA jobs
to control the access to the RDS service, we are forced to create a new
network only to add the security network.

To define the new networks we will:
 - Pick a range of 10 IPs from the original networks.
 - Mark this range as reserved in the original networks.
 - Create the new networks `cfX_rds` with the same configuration than the
   original ones: DNS, router, range...
 - The new network has all the IPs reserved except the desired
   range of 10 IPs.

The new networks will include the security_groups entry, including the
same value than the default_security_groups defined in the bosh manifest
and the new security group for RDS clients.

We explicitelly add each one of the ranges in the definitions to make it
more explicit and easy to read. The dynamic IP ranges are calculated
automatically by BOSH from reserved and static, but we add a `dynamic` entry
to make it more obvious.

In other to set the security groups for the API and UAA vms, we need
to use the previously defined networks cf1_rds, cf2_rds. etc in the job
definition.