Setup a default IAM profile for all deployed VMs
When setting up BOSH blobstore on S3, all the agents will access the
S3 bucket directly to grab the binaries.

As we want to use IAM profiles to pass the credentials, wWe created an IAM
profile to be set as a default called `bosh-managed`.

We need to configure bosh to set this profile as default for all VMs.

Note: Any other IAM profile created must have the same policies to be
able to access the bosh blobstore.