Init database pipeline tasks
The previous postgres job was doing some work to initialise
the databases: create roles, create databases, enable extensions.

Most of this work should have been done with the postgres terraform
provider, but it has bugs and can't destroy the resources properly:
https://github.com/hashicorp/terraform/issues/5340

When this is fixed, we can move the create role and db tasks to
terraform, but we still need to enable extensions.