Update deployer garden linux release to 334
Also configure 3GB container cache. Only after this threshold is
crossed will garden start doing "graph cleanups" for containers. This
means that containers that are used, but then deleted, will not be
immediately removed, as it was before, if they fit within 3GB cache,
which is plenty given we use alpine mostly and strive to have small
containers.