Update pipeline to use new cf secret generator.
This now inits the secrets with a zero bytes file in the init step, and
then gets and puts the resource in the generate step. This allows us to
pin the secrets in the terraform job to the generate job's output.

Now that the script is no longer using mkpasswd, there's no need to skip
that test on OS X