Added job to deploy graphite-nozzle
graphite-nozzle connects to the traffic controller firehose service and
sends all the metrics to a given graphite server.

We deploy it as a CF application as a short-term solution, before we do
not implement a proper bosh-release. Because that, there is no logic to
avoid redeploying the app every time the pipeline is run.

We will deploy the patched version
https://github.com/alphagov/graphite-nozzle/tree/106191208-deaggregated_names
(branch checked out in the resource) which adds prefix with the job name and
index in every metric.

We rely on several settings from the manifest.yml and cf-secrets.yml,
which are read using the val_from_yaml.rb script.

The application is deployed using a generated manifest.yml file, which
includes several options and the environment variables.