Upgrade vagrant kernel version
The latest version of concourse includes a [release that makes use of](https://github.com/cloudfoundry-incubator/garden-linux/commit/36e1270b4b5586efce8941499dad635b5d5f5148) the
[execveat syscall](http://man7.org/linux/man-pages/man2/execveat.2.html) which is not supported on kernel version 3.13 which the
concourse ami uses.

To workaround this, we can upgrade the kernel. To do
this we must reboot the machine. I opted to do the reboot during the
bootstrap process and wait for the machine to reboot because it will
keep the current behaviour of having an ssh tunnel running. To avoid
waiting every time the script is run, which is useful to update
pipelines or restart the ssh tunnel, this will only happen if a file is
present - the destroy script removes this file, so it should only exist
if the script is run or someone manually creates the file. The script
doesnt match against an exact kernel version, as it is likely the kernel
version will be updated soon - instead it looks for a 3.19 kernel after
installing the vivid kernel package and uses that to replace the current
kernel version.