Bootstrap concourse - configure external URL
External URL in the AMI image is hardcoded to some IP which was likely
used during AMI build, but is not valid anymore. This prevents logins
via browser (it redirects to this IP). Replace this with empty quotes
instead, which redirects back to whichever URL you use to access it.