Make the mount persist across reboots
Because we now need to reboot for the kernel update to take effect, we
need to make sure the mount persists across reboots, which it wasnt
doing before. Printf needs to be used to expand the tab characters, and
grep is used to ensure we dont write the same line multiple times.