Trigger everything from pipeline-trigger.
Most things in the creation pipeline were already doing this, the only
outstanding things in this pipeline were the tests. This updates these
to trigger from pipeline-trigger, and nothing else.

For the destroy-cloudfoundry and destroy-microbosh pipelines, this
updates them to use a similar pattern with a run starting with an init
job that bumps the pipeline-trigger, and everything else triggering from
that.