Set base retention and aggregation for all metrics
Remove example and use the following policy:
   * Keep 1 day metrics in 10s interval
   * Keep 1 week metrics in 1 min interval
   * Keep 1 months metrics in 15 min interval
   * Keep 1 year metrics in 1 hour interval
   * Keep 5 year metrics in 1 day interval

Given these settings, each metric will take up 343kb (See
https://gist.github.com/jjmaestro/5774063). As cloud foundry produces
thousands of metrics, this should be reviewed at a later date.

Apply to all metrics this aggregation policy:
when the data points are aggregated because of the retention rules above,
we average the values if at least half of the data points are not null.
If not, the aggregate is null.