Pin grafana and graphite release using tag_filter
We want to pin the releases of grafana and graphite to a specific
commit. But concourse git resource does not support pinning versions,
until the generic pin solution is implemented.

As a workaround, we fork and tag the repository with tags called
`commit-<hash>`, and we use the `tag_filter` feature from the git
resource.

[1] https://github.com/concourse/concourse/issues/256