Add job to autoupdate create-cloudfoundry pipeline
The job will download the fly command directly from the concourse server.
Note: I found some issues using a different fly command version than the
server, so that the command `fly sync` did not work. It is better to
always download it from the server using curl or wget.

Then login in concourse so that we can use the same scripts to update the
pipeline itself.

We manually populate each variable required as environment variables, so
that the concourse/scripts/pipelines-bosh-cloudfoundry.sh script can
regenerate and pass the variables again when rendering and uploading
the pipeline manifest.