Add self update to create-cloudfoundry pipeline
This job will call the same make target for the given environment tasks
using the `$MAKEFILE_ENV_TARGET` variable.

The intention of the variable `$MAKEFILE_ENV_TARGET` is exclusively
identify the right make target to pass in the self update job, and it
should not be used to parametrise any environment specific behaviour.

In the non-development environments, all the variables values will be
determined by the committed code and the variables set in
entry point in the Makefile.

The task makes use of the existing logic to download the `fly` command
and login in the concourse environment. The only non reproducible variable
is the `CONCOURSE_ATC_PASSWORD` which we retrieve from the concourse
manifest.