Allow specify which pipelines update in concourse
When doing the auto-update pipeline, each pipeline will update only
itself. But the scripts that upload the pipelines update several in
the same script.

We add logic to read a variable PIPELINES_TO_UPDATE so that we can
specify only one pipeline to update.

We also remove the dedicated script `pipelines-failure-testing.sh` which
create the `failure-testing` pipeline, as it is related to cloudfoundry,
and all the variables and logic is the same. This allows us
define a unique point where whe decide the pipelines to update.