Remove TCP/9300 from sg.elastic_master_elb
During a discussion about the pull request we decided to remove TCP/9300
from the ELB because it is only used by other Elasticsearch nodes (which
don't use the ELB) and Java clients (of which we don't have any).

- https://www.elastic.co/guide/en/elasticsearch/reference/current/modules-transport.html

It was rebased out of the commit which added the ELB, but we forgot to
remove it from the security group.