Only sync fly when it's more than a day old.
`fly sync` doenloads the fly binary from concourse every time, which can
be time consuming when it's being called frequently.

This adds a conditional around it to only update if it's more than a day
old.

Note: This is using `find` to determine the file age as it's the best
option that's portable across BSD (OS X) and Linux. Some other things
that I tried:

- `stat` has different options between platforms, and there's no common option
  that does what's needed.
- `date -r` is only available on Linux.