Parametrise environment by environment class
Currently we deploy the following different environments classes:

 - dev: For development of the platform configuration
 - ci: For continuous integration
 - stage: For staging testing
 - prod: For end user usage

each one of these environments will have different configuration sets.
In order to simplify setting this variables, specially in the
self-update pipeline job, we will define a global variable to
identify the current environment class.

That the variable ENVIRONMENT_CLASS, which will determine
the specific set of variables per environment. The Makefile will set this
variable and propagate it across the scripts, and it will set in the
pipelines as the parameter `environment_class`.

Specifically, this variable currently determines the AWS account to use and
if the self-update pipeline feature is enabled or not, which by default not
enabled in dev.

A environment will be uniquely determined by its "environment class",
"deploy name" and "branch".