Use FQDN in concourse dns name output.
This will avoid the trailing . in the hostnames used for the
deployer-concourse. This is also the more correct attribute to use
according to the docs[1] as the name can be specified relative to the
zone name.

This output is only currently used in the concourse manifest to specify
the external_url for atc, which is used when redirecting after auth.

[1]https://www.terraform.io/docs/providers/aws/r/route53_record.html