Cloud-controller-ng logging level
Cloud Controller is logging in "debug2" mode and generating a lot of
unnecessary noise. This commit reduces logging level to info
so we can get more sensible level that will still allow us to identify
platform problems. We left db log level at debug as this one sets a
verbosity level at which DB logs - it's not filtering. If we want to see
all SQL queries we have to change cc log level back to debug.