Explicitly add protocol into CATS api property
There is issue with v3 CATS tests where they don't add protocol in to
the API endpoint and thus end up trying to connect to http, which is
closed. Add the protocol into CATS api endpoint explicitly until the
issue is fixed and released upstream.

This is a standalone commit so that it can be easily reverted when we
use the fixed version.