Generate a SSH key for concourse git repos
In order to use the pool resource to lock pipelines, which uses a
git based backend, we need to generate a new SSH key pair.

This keys will be different than the ones uses on the servers because
security reasons, as it will be passed to the pipelines as parameter.

We use the same code to generate the one unique key per environment only
once.

The key will be called git_id_rsa.