Add pool resource and download and pass git_id_rsa
Add the pool resource to the create-bosh-cloudfoundry pipeline.

We need download the `git_id_rsa` SSH key and the terraform concourse
tfstate from the state S3 bucket pass the key and the generated
codecommit url as variables for the pipelines.

For that, we rely on AWS Cli, which is added as a dependency to
be installed to locally for the initial deployment.

AWS cli is also a dependency for the self-update-pipeline container
which was extended here:
https://github.com/alphagov/paas-docker-cloudfoundry-tools/pull/43

NB: We use bash variable substituion as in `${git_id_rsa//$'\n'/$'\n'  }`
to avoid the shellcheck warning SC2001 [1]

[1] https://github.com/koalaman/shellcheck/wiki/SC2001