Test for closed http
Add a test that verifies that http is closed for both API and apps.
We have picked 11 second timeout as not too long to delay tests for too
much and also not too short to be caused by random events like concourse
being too slow (when it launches 5 parallel tests), internet interfering
with TCP packets (between councourse and ELB) or even ELB having some
random hiccup.