Create IAM user to access codecommit git repo
In order to access a codecommit git repository you must use a IAM user with
the right permissions.

In this case we create a user called `git-${DEPLOY_ENV}` which matches our
the policy which allows the creation of users called `git-*`

In order to grant permissions to access codecommit git repositories for
the pool resource, we add the user to the group concourse-pool-git-rw.
This group has a policy associated to be able to access codecommit git
repos called `concourse-pool-*`.

Note that we cannot use the resource `aws_iam_group_membership` from
terraform, as it would wipe all the members of the group until this
feature request is not solved: _https://github.com/hashicorp/terraform/issues/5778

Because that we add a new task after terraform that calls the awscli to add
the user to the group. This is fine because the user will be removed
automatically from the group if it is deleted.

We prefer this alternative rather than doing local_exec to avoid start a
tendency implement call outs to awscli in terraform.