Upload git ssh public key to git user
We want to upload the generated SSH key to the created git user.

We use the resource `aws_iam_user_ssh_key` from terraform, implemented
in _https://github.com/hashicorp/terraform/issues/5744 so this requires
an updated version of terraform.

We retrieve the generated key as a s3 resource and pass it to terraform
as a TF_VAR_ variable.

After the key is uploaded, AWS assigns it a unique id which must be used
as user for SSH when connecting to the codecommit git repositories.

We render and output a full url with that ssh_key_id and the ssh url of the
repository. Note that the url must be in the ssh:// format, or
codecommit will drop the git connections.