Remove our own simple_app_test
This was added when we introduced the custom acceptance test suite alongside
the existing Cloud Foundry acceptance tests. It duplicates functionality in
the CF tests and was a placeholder until we added other tests to our custom
suite.

e18e80b adds our first real custom acceptance test so we no longer need to
have this. We need to keep `tests/acceptance-tests/example_apps` though
because it's used by that new HTTP test.