Workaround terraform codecommit issue
Terraform codecommit resource[1] has a bug managing the default_branch
property.

Although this property is optional, terraform will try to manage it.

If defined, codecommit will fail during creation:

  BranchDoesNotExistException: refs/heads/master does not exist

Meanwhile if default_branch is defined, codecommit it will fail as soon
as a branch is pushed to the repository. The reason is that AWS enforces
having at least one branch as default branch, failing when terraform is
trying to set it empty:

  * aws_codecommit_repository.concourse-pool: Error Updating Default
  * Branch for CodeCommit Repository: InvalidParameter: 1 validation
    * errors:
     - field too short, minimum length 1: DefaultBranchName

This will be reported upstream, but meanwhile we will workaround it by
passing the default branch as a variable, and quering it using awscli.
If the respository does not exist or does not have default branch, the
variable will be set to empty string "".