Refactor certificate generation
Rather than have certificates generated in 3 separate places for
metrics, logsearch, and cloudfoundry certificates, 2 places using
openssl and 1 place using certstrap, consolidate to 1 task which uses
certstrap to generate certficates. This does mean the metrics and
logsearch certificates are now signed by the bosh CA rather than being
self signed, but as this is for internal use only, this should not have
any impact. The generation has also been moved earlier in the pipeline,
so that the certficates can be used by terraform to upload iam server
certificates and associate with the elb.