Added ALERT_EMAIL_ADDRESS variable
This variable is used in the "TO" and "FROM" of the alert email. It is
also used to control whether the continuous smoke tests should be
triggered at all.

We use `sed` because fly only substitutes strings so it adds double quotes
to any value and it breaks for booleans.