Add script to send email notification
To simplify the usage of AWS Simple Email Service we have used the
AWS CLI tool. We did not use the Concourse email resource type as it
does not support instance profiles.

This script creates an email with details from the smoke tests (link to
Concourse, last commit, smoke tests log) and send it using the AWS CLI:
http://docs.aws.amazon.com/cli/latest/reference/ses/send-email.html

The message is formatted as HTML. The content from git and log is
crudely processed so it can be displayed as html and still can fit
into the json payload (basically we remove double quotes and line
endings).