Download and allow pingdom probe IPs to router
Download the IPs from the pingdom feed with the list of IPs[1], and
add them to security groups to allow access to the routers.

Due the limits in AWS related to security groups in our account,
we cannot put more than 50 rules in each security group. Each CIDR will
count as a rule. To workaround that, we create 3x security groups
for the pingdom groups, splitting the list of IPs in batches in
shell script in the pipeline.

We do not use `count` in terraform, but instead we repeat the SG
definition  because we must use one different variable per group of IPs.
This is because terraform does not support externally define variables
as hashes.

[1] https://help.pingdom.com/hc/en-us/articles/203682601-How-to-get-all-Pingdom-probes-public-IP-addresses