Make concourse publicly_viewable.
This allows read-only access to the concourse UI, but not to the output
of job runs (unless the job itself is defined as public). See
http://bosh.io/jobs/atc?source=github.com/concourse/concourse&version=0.74.0#p=atc.publicly_viewable
for more details.

This is necessary to allow us to have the pipelines on the monitoring
screens safely.

With this in place, access is still locked down by IP.