Make travis build fail early on download error
Travis doesn't run these script blocks with `set -e`, so if one line
errors, it will continue, and as long as the last line exits zero, the
build will continue, only to fail later. Adding `set -e` makes it fail
fast and therefore makes it easier to see what the problem is.