Added healthcheck application to pipeline
We need a basic static application to provide something
to healthcheck via 3rd party services such as Pingdom.

There was an existing directory for example applications
which we have refactored, as it was just for acceptance
tests. Now this directory can store example applications
for any purpose.

A task was created which uses CF CLI to push the static
application.