Scripts for creating Pingdom checks with Terraform
We wanted to simplify the automation of creating Pingdom
checks by utilising a Terraform Pingdom provider.
We have used: https://github.com/russellcardullo/terraform-provider-pingdom
Binaries for this provider (for Linux and OSX) have been
included in this repo.

A Makefile target has been created to make use of the
dynamically created healthcheck application URLs.

The makefile target `pingdom` must be run after an environment has been
deployed - e.g

```
make dev pingdom DEPLOY_ENV=<env>
```

These must be run outside of the ci pipeline, and on the deploying users
laptop as they rely on getting pingdom authentication information from
the paas-pass repo, which is hosted on github enterprise and is
difficult to give concourse access to.