Add API SG group and dedicated ELB for cc
Add a dedicated ELB for the Cloud Foundry CloudController and a
CF API security group to allow access to the API endpoint to only
the office IPs, the NAT nodes and concourse.

The new ELB will use the same configuration than the go-router ELB, but
use the API nodes as backend. Cloud Controller listens in port 9022,
without SSL. Health check can be done in the endpoint /info.

CC does not require WebSockets, so we can enable the HTTP protocol in
the ELB configuration.