Remove uaa.proxy.servers property
This added the gorouter IPs to the list of IPs that were trusted when
evaluating the X-Forwarded-For header. This was actually unnecessary
because the defualt list includes all RFC1918 addresses[1].

Remove this property because the traffic no longer comes via the
gorouter, so this would be misleading. There's no need to add entries
for the ELBs because these will use RFC1918 addresses.

[1]http://bosh.io/jobs/uaa?source=github.com/cloudfoundry/cf-release&version=228#p=uaa.proxy_ips_regex