Introduce Pipecleaner & hook into Travis tests
Pipecleaner is a small tool I wrote to help spot concourse pipeline
errors quickly before we actually run them in concourse. It was written
during another story to help with refactoring pipelines and made the
whole process substantially faster.

It can check for the following issues:

* Resources being used in a job that have not been defined in the
  `resources:` block. (Fatal)
* Resources that have been defined in the `resources:` block and are not
  used in the pipeline. (Warning)
* Resources being used in an `input:` block that have not been
  `get:`-ted. (Fatal)
* Resources that have been `get:`-ted and are not used in the job.
  (Warning)
* `output:`s that are not used later in the job. (Warning)

By default it will exit with a nonzero exit code for any Fatal errors,
and will exit with a nonzero code for Warnings if you pass the
`--fatal-warnings` flag.

If some checks are not desired (e.g. unused outputs) they can be
disabled with the `--ignore-types` flag.

Travis is configured to test all concourse pipelines by default, but
will only fail the tests on Fatal, not Warning level errors.